backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.4
pydantic-settings==2.9.1
numpy>=1.24.0
pillow>=10.0.0
aiofiles>=23.0.0
openai>=1.30.0
google-genai>=1.0.0
